+++ plugins/model-apps/scripts/lib/app-spec.js
@@ -1277,5 +1277,6 @@ module.exports = {
   prefixedRelationshipName,
   manyToManyFor,
   manyToManySchemaName,
+  isSafeHttpUrl,
   CHART_TYPES,
 };